# docs/swagger.yaml
definitions:
  dberrors.ConflictError:
    type: object
  dberrors.ZeroRowsAffectedError:
    type: object
...
  models.Vendor:
    properties:
      address:
        type: string
      contact:
        type: string
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      vendor_id:
        type: string
    type: object
  server.Response:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /vendors:
    delete:
      consumes:
      - application/json
      description: Delete a vendor from the database by its ID
      parameters:
      - description: Vendor ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dberrors.ZeroRowsAffectedError'
      summary: Delete a vendor
      tags:
      - vendors
    get:
      consumes:
      - application/json
      description: Retrieve a list of all vendors with optional pagination
      parameters:
      - description: Page index for pagination
        in: query
        name: pageindex
        type: string
      - description: Page size for pagination
        in: query
        name: pagesize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Vendor'
            type: array
      summary: Get all vendors
      tags:
      - vendors
    post:
      consumes:
      - application/json
      description: Create a new vendor in the database
      parameters:
      - description: Vendor to add
        in: body
        name: vendor
        required: true
        schema:
          $ref: '#/definitions/models.Vendor'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vendor'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dberrors.ConflictError'
        "415":
          description: Unsupported Media Type
          schema:
            type: string
      summary: Add a new vendor
      tags:
      - vendors
  /vendors/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a vendor by its ID
      parameters:
      - description: Vendor ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Vendor'
      summary: Get vendor by ID
      tags:
      - vendors
  /vendors/{vendor_id}:
    put:
      consumes:
      - application/json
      description: Update a vendor's details by providing its ID
      parameters:
      - description: Vendor ID
        in: path
        name: vendor_id
        required: true
        type: string
      - description: Updated vendor data
        in: body
        name: vendor
        required: true
        schema:
          $ref: '#/definitions/models.Vendor'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Vendor'
        "400":
          description: Bad Request
          schema:
            type: string
        "415":
          description: Unsupported Media Type
          schema:
            type: string
      summary: Update an existing vendor
      tags:
      - vendors
swagger: "2.0"